+++ CHANGELOG.md
@@ -10,8 +10,8 @@ script to proceed.
 - Move `temboard-migratedb` as `temboard migratedb` subcommand.
 - Move `temboard-agent-register` as `temboard-agent register` subcommand.
 - New subcommands `temboard query-agent`, `temboard routes`, `temboard
-  runtask`, `temboard web`, `temboard-agent routes`, `temboard-agent runtask`
-  and `temboard-agent web` for debugging.
+  runtask`, `temboard schedule`, `temboard web`, `temboard-agent routes`,
+  `temboard-agent runtask` and `temboard-agent web` for debugging.
 - temBoard agent auto_configure.sh does not configure file logging anymore.
   temBoard UI & agent package does not ship logrotate file anymore. temBoard UI
   & agent can still log to file if you configure it to do so.
@@ -24,6 +24,7 @@ script to proceed.
 - Post install script of temBoard UI RPM does not execute auto_configure.sh.
 - temBoard UI RPM does not create `temboard` UNIX user. Use auto_configure.sh
   instead.
+- Fast collect of monitoring and statements metrics upon agent registration.
 - temBoard UI dropped push-metrics handler from pre-6.0 push metric collect.
 - Python 2.7 and 3.5 support is deprecated and will be removed in next major
   release.